user login functionality

+++ base/views.py
@@ -1,11 +1,44 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
 
 
 # Create your views here.
+
+def login_page(request):
+
+  if request.user.is_authenticated:
+    return redirect('home')
+
+  if request.method == "POST":
+    username = request.POST.get('username')
+    password = request.POST.get('password')
+
+    try:
+      user = User.objects.get(username=username)
+    except:
+      messages.error(request, "User does not exist.")
+      return redirect('login')
+    
+    user = authenticate(request, username=username, password= password)
+
+    if user is not None:
+      login(request, user)
+      return redirect('home')
+    else:
+      messages.error(request, "Incorrect username or password.")
+  context={}
+  return render(request, 'base/login_register.html',context)
+
+def logout_user(request):
+  logout(request)
+  return redirect("home")
 
 def home(request):
   q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -28,6 +61,7 @@ def room(request, pk):
     
   return render(request, 'base/room.html', context)
 
+@login_required(login_url="login")
 def create_room(request):
   form = RoomForm()
 
@@ -40,9 +74,14 @@ def create_room(request):
   context = {"form": form}
   return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url="login")
 def update_room(request, pk):
   room = Room.objects.get(id=pk)
   form = RoomForm(instance=room)
+
+  if request.user != room.host:
+    return HttpResponse("You are not allowed!!")
 
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
@@ -53,8 +92,14 @@ def update_room(request, pk):
   context = {"form": form}
   return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url="login")
 def delete_room(request, pk):
   room = Room.objects.get(id=pk)
+
+  if request.user != room.host:
+    return HttpResponse("You are not allowed!!")
+  
   if request.method == "POST":
     room.delete()
     return redirect('home')
